refactor(backends): decompose claude.py; fix _validate_codex_mcp_inventory re-export
Extract ClaudeStreamParser + ClaudeResultParser to _claude_parse.py and
ClaudeSessionLocator to _claude_session_locator.py. Re-export from
claude.py so the public surface is unchanged.

Also re-add _validate_codex_mcp_inventory to the _codex_probes re-export
block (it was dropped by the linter since the import was not used
inside codex.py itself; evidence_reader imports it via the canonical
codex path).

Refs: #4664
