+++ src/autoskillit/execution/backends/codex.py
@@ -135,6 +135,7 @@ from autoskillit.execution.backends._codex_prelaunch import codex_prelaunch_tran
 # can keep importing them from the canonical codex module path.
 from autoskillit.execution.backends._codex_probes import (
     _BoundedProbeResult,  # noqa: F401
+    _validate_codex_mcp_inventory,
     _validate_global_codex_home,
     _validate_inert_rollout_paths,
     _validate_mcp_probe,
